Handling more exceptions

+++ BackendTask/UserManager.cs
@@ -28,55 +28,78 @@ public class UserManager
     {
         var currUser = user;
         currUser.id = CreateId();
-        currUser.updated_at = DateTime.Now;
-        currUser.created_at = DateTime.Now;
-        var results = IsValidated(currUser);
-        if (results == null)
+        if (currUser.id == 0)
         {
-
-            Console.WriteLine(currUser.id);
-
-            return _userDatabaseRepository.CreateUser(currUser);
+            return "Cannot create id, maybe all ids are already taken";
         }
-        else 
+        else
         {
-            string ErrorString = "";
-            foreach (var error in results)
+            currUser.updated_at = DateTime.Now;
+            currUser.created_at = DateTime.Now;
+            var results = IsValidated(currUser);
+            if (results == null)
             {
-                ErrorString += error.ErrorMessage + System.Environment.NewLine;
+
+                Console.WriteLine(currUser.id);
+
+                return _userDatabaseRepository.CreateUser(currUser);
             }
-            return ErrorString;
+            else
+            {
+                string ErrorString = "";
+                foreach (var error in results)
+                {
+                    ErrorString += error.ErrorMessage + System.Environment.NewLine;
+                }
+                return ErrorString;
+            }
         }
+       
         
     }
     public string DeleteUser(int id)
     {
-        return _userDatabaseRepository.DeleteUser(id);
+        var getedUser = _userDatabaseRepository.GetUserById(id);
+        if (getedUser == null)
+        {
+            return "No user with such id!";
+        }
+        else
+        {
+            return _userDatabaseRepository.DeleteUser(id);
+        }
     }
     public string UpdateUser(ValidatedUserDto user)
     {
         var curruser = user;
-        curruser.created_at = _userDatabaseRepository.GetUserById(curruser.id).created_at; 
-        curruser.updated_at= DateTime.Now;
-        var results = IsValidated(curruser);
-        if (results == null)
+        var getedUser = _userDatabaseRepository.GetUserById(curruser.id);
+        if (getedUser == null)
         {
-
-
-            Console.WriteLine(curruser.id);
-
-            return _userDatabaseRepository.UpdateUser(curruser);
+            return "No user with such id!";
         }
         else
         {
-            string ErrorString = "";
-            foreach (var error in results)
+            curruser.created_at = getedUser.created_at;
+            curruser.updated_at = DateTime.Now;
+            var results = IsValidated(curruser);
+            if (results == null)
             {
-                ErrorString += error.ErrorMessage+System.Environment.NewLine;
-               
+                Console.WriteLine(curruser.id);
+                return _userDatabaseRepository.UpdateUser(curruser);
             }
-            return ErrorString;
+            else
+            {
+                string ErrorString = "";
+                foreach (var error in results)
+                {
+                    ErrorString += error.ErrorMessage + System.Environment.NewLine;
+
+                }
+                return ErrorString;
+            }
         }
+           
+       
     }
     public ValidatedUserDto GetUser(int id)
     {
@@ -92,16 +115,22 @@ public class UserManager
     }
     private int CreateId()
     {
+        int c = 0;
         Random rnd = new Random();
         var buffer = new byte[sizeof(int)];
         rnd.NextBytes(buffer);
         var id= Math.Abs(BitConverter.ToInt32(buffer, 0));
         while(_userDatabaseRepository.GetUserById(id)!=null)
         {
+            
             var buffer1 = new byte[sizeof(int)];
             rnd.NextBytes(buffer1);
             id = Math.Abs(BitConverter.ToInt32(buffer1, 0));
-            
+            if (c > 10000)
+            {
+                return 0;
+            }
+            c++;
         }
         return id;
     }
